backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
python-jose==3.3.0
requests==2.32.3
pydantic==2.8.2
email-validator==2.1.0
